+++ test/unicode_set_test.exs
@@ -6,6 +6,7 @@ defmodule UnicodeSetTest do
   doctest Property
   doctest Parser
   doctest Sigil
+  doctest Unicode.Regex
 
   test "set intersection when one list is a true subset of another" do
     l = Unicode.GeneralCategory.get(:L)
@@ -199,5 +200,37 @@ defmodule UnicodeSetTest do
   test "creating unicode classes for regex" do
     assert Unicode.Set.character_class("[{HZ}]") == "{HZ}"
     assert Unicode.Set.character_class("[[:Lu:]&[AB{HZ}]]") == "A-B"
+  end
+
+  test "parsing invalid unicode classes for regex" do
+    assert Unicode.Set.character_class("[:ZZZ:]") ==
+    {:error,
+      {Unicode.Set.ParseError,
+        "Unable to parse \"[:ZZZ:]\". The unicode script, category or property \"zzz\" is not known."}}
+  end
+
+  test "parsing invalid regex" do
+    assert Unicode.Regex.compile("[:ZZZ:]") ==
+    {:error,
+      {Unicode.Set.ParseError,
+        "Unable to parse \"[:ZZZ:]\". The unicode script, category or property \"zzz\" is not known."}}
+  end
+
+  test "parsing an invalid unicode set returns the right error" do
+    assert Unicode.Set.parse("[:ZZZZ:]") ==
+    {:error,
+     {Unicode.Set.ParseError,
+      "Unable to parse \"[:ZZZZ:]\". The unicode script, category or property \"zzzz\" is not known."}}
+  end
+
+  test "parsing perl and posix positive and negative regex" do
+    assert Unicode.Regex.compile("[:Zs:]") ==
+      {:ok, ~r/[\x{20}\x{A0}\x{1680}\x{2000}-\x{200A}\x{202F}\x{205F}\x{3000}]/u}
+    assert Unicode.Regex.compile("[:^Zs:]") ==
+      {:ok, ~r/[^\x{20}\x{A0}\x{1680}\x{2000}-\x{200A}\x{202F}\x{205F}\x{3000}]/u}
+    assert Unicode.Regex.compile("\\P{Zs}") ==
+      {:ok, ~r/[^\x{20}\x{A0}\x{1680}\x{2000}-\x{200A}\x{202F}\x{205F}\x{3000}]/u}
+    assert Unicode.Regex.compile("\\p{Zs}") ==
+      {:ok, ~r/[\x{20}\x{A0}\x{1680}\x{2000}-\x{200A}\x{202F}\x{205F}\x{3000}]/u}
   end
 end